# services/customer-service/swagger.yaml
openapi: 3.0.0
info:
  title: CRM Service API
  description: API complète de gestion CRM - Clients, Contacts, Contrats, Opportunités
  version: 1.0.0
  contact:
    name: Support CRM
    email: support@crm-service.com
servers:
  - url: http://localhost:4007
    description: Serveur de développement
  - url: https://api.crm-service.com
    description: Serveur de production
...
paths:
...
  /api/clients:
    get:
      tags:
        - Clients
      summary: Lister tous les clients
      description: Récupère la liste des clients avec pagination et filtres
      parameters:
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filtre par statut
          required: false
          schema:
            type: string
            enum: [PROSPECT, ACTIF, INACTIF, SUSPENDU, ARCHIVE, LEAD_CHAUD, LEAD_FROID]
        - name: typeClientId
          in: query
          description: Filtre par type de client
          required: false
          schema:
            type: string
        - name: search
          in: query
          description: Recherche textuelle
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Liste des clients récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientsListResponse'
        '401':
          description: Non authentifié
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Clients
      summary: Créer un nouveau client
      description: Crée un nouveau client dans le système
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '400':
          description: Données invalides
        '401':
          description: Non authentifié
        '403':
          description: Non autorisé (rôle insuffisant)
        '409':
          description: Client avec cet email existe déjà
        '500':
          description: Erreur interne du serveur
...
  /api/contacts:
    get:
      tags:
        - Contacts
      summary: Lister les contacts
      description: Récupère la liste des contacts avec filtres
      parameters:
        - name: clientId
          in: query
          description: Filtre par client
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Filtre par type de contact
          required: false
          schema:
            type: string
            enum: [COMMERCIAL, TECHNIQUE, COMPTABLE, DIRECTION, SUPPORT, AUTRE]
        - name: statut
          in: query
          description: Filtre par statut
          required: false
          schema:
            type: string
            enum: [ACTIF, INACTIF, PARTI]
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des contacts récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Contacts
      summary: Créer un contact
      description: Crée un nouveau contact pour un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '409':
          description: Contact avec cet email existe déjà pour ce client
        '500':
          description: Erreur interne du serveur
...
  /api/contrats:
    get:
      tags:
        - Contrats
      summary: Lister les contrats
      description: Récupère la liste des contrats avec filtres
      parameters:
        - name: clientId
          in: query
          description: Filtre par client
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filtre par statut
          required: false
          schema:
            type: string
            enum: [BROUILLON, EN_ATTENTE_SIGNATURE, ACTIF, SUSPENDU, RESILIE, TERMINE, EN_RENOUVELLEMENT]
        - name: typeContrat
          in: query
          description: Filtre par type de contrat
          required: false
          schema:
            type: string
            enum: [MAINTENANCE, SERVICE, PRODUIT, PARTENARIAT, ABONNEMENT, FORFAIT, CONSULTING, FORMATION, LICENCE, SAAS]
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des contrats récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratsListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Contrats
      summary: Créer un contrat
      description: Crée un nouveau contrat pour un client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContratRequest'
      responses:
        '201':
          description: Contrat créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '409':
          description: Numéro de contrat déjà existant
        '500':
          description: Erreur interne du serveur
...
  /api/adresses:
    get:
      tags:
        - Adresses
      summary: Lister les adresses
      description: Récupère les adresses d'un client
      parameters:
        - name: clientId
          in: query
          description: ID du client
          required: true
          schema:
            type: string
        - name: typeAdresse
          in: query
          description: Filtre par type d'adresse
          required: false
          schema:
            type: string
            enum: [FACTURATION, LIVRAISON, SIEGE_SOCIAL, ETABLISSEMENT, CORRESPONDANCE]
      responses:
        '200':
          description: Liste des adresses récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdressesListResponse'
        '400':
          description: clientId manquant
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Adresses
      summary: Créer une adresse
      description: Crée une nouvelle adresse pour un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdresseRequest'
      responses:
        '201':
          description: Adresse créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdresseResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '500':
          description: Erreur interne du serveur
...
  /api/interactions:
    get:
      tags:
        - Interactions
      summary: Lister les interactions
      description: Récupère la liste des interactions avec filtres
      parameters:
        - name: clientId
          in: query
          description: Filtre par client
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Filtre par type d'interaction
          required: false
          schema:
            type: string
            enum: [APPEL, EMAIL, REUNION, VISITE, SUPPORT, COMMERCIAL, TECHNIQUE, FORMATION, DEMONSTRATION, PRESENTATION, NEGOCIATION]
        - name: canal
          in: query
          description: Filtre par canal
          required: false
          schema:
            type: string
            enum: [TELEPHONE, EMAIL, EN_PERSONNE, VIDEO, CHAT, RESEAUX_SOCIAUX, PORTAL_CLIENT, MOBILE]
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des interactions récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionsListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Interactions
      summary: Créer une interaction
      description: Crée une nouvelle interaction pour un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInteractionRequest'
      responses:
        '201':
          description: Interaction créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '500':
          description: Erreur interne du serveur
...
  /api/documents:
    get:
      tags:
        - Documents
      summary: Lister les documents
      description: Récupère la liste des documents avec filtres
      parameters:
        - name: clientId
          in: query
          description: Filtre par client
          required: false
          schema:
            type: string
        - name: typeDocument
          in: query
          description: Filtre par type de document
          required: false
          schema:
            type: string
            enum: [CONTRAT, FACTURE, DEVIS, AVENANT, KYC, LEGAL, TECHNIQUE, COMMERCIAL, ADMINISTRATIF, FINANCIER, RAPPORT]
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des documents récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Documents
      summary: Uploader un document
      description: Upload un nouveau document pour un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document uploadé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '500':
          description: Erreur interne du serveur
...
  /api/opportunites:
    get:
      tags:
        - Opportunités
      summary: Lister les opportunités
      description: Récupère la liste des opportunités avec filtres
      parameters:
        - name: clientId
          in: query
          description: Filtre par client
          required: false
          schema:
            type: string
        - name: etape
          in: query
          description: Filtre par étape
          required: false
          schema:
            type: string
            enum: [PROSPECTION, QUALIFICATION, PROPOSITION, NEGOCIATION, FINALISATION]
        - name: statut
          in: query
          description: Filtre par statut
          required: false
          schema:
            type: string
            enum: [OUVERTE, GAGNEE, PERDUE, MISE_EN_ATTENTE]
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des opportunités récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunitesListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Opportunités
      summary: Créer une opportunité
      description: Crée une nouvelle opportunité commerciale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportuniteRequest'
      responses:
        '201':
          description: Opportunité créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportuniteResponse'
        '400':
          description: Données invalides
        '404':
          description: Client non trouvé
        '500':
          description: Erreur interne du serveur
...
  /api/type-clients:
    get:
      tags:
        - Types Clients
      summary: Lister les types de clients
      description: Récupère la liste des types de clients
      parameters:
        - name: isActive
          in: query
          description: Filtre par statut actif
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Liste des types récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeClientsListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Types Clients
      summary: Créer un type de client
      description: Crée un nouveau type de client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTypeClientRequest'
      responses:
        '201':
          description: Type créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeClientResponse'
        '400':
          description: Données invalides
        '409':
          description: Code déjà existant
        '500':
          description: Erreur interne du serveur
...
  /api/secteurs:
    get:
      tags:
        - Secteurs
      summary: Lister les secteurs
      description: Récupère la liste des secteurs d'activité
      parameters:
        - name: parentId
          in: query
          description: Filtre par secteur parent
          required: false
          schema:
            type: string
        - name: niveau
          in: query
          description: Filtre par niveau
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Liste des secteurs récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecteursListResponse'
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - Secteurs
      summary: Créer un secteur
      description: Crée un nouveau secteur d'activité
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecteurRequest'
      responses:
        '201':
          description: Secteur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecteurResponse'
        '400':
          description: Données invalides
        '404':
          description: Secteur parent non trouvé
        '409':
          description: Code NAF déjà existant
        '500':
          description: Erreur interne du serveur
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateClientRequest:
      type: object
      required:
        - nom
        - email
        - typeClientId
      properties:
        nom:
          type: string
          example: TechnoSoft Solutions
        raisonSociale:
          type: string
          example: TechnoSoft Solutions SAS
        email:
          type: string
          format: email
          example: contact@technosoft.fr
        telephone:
          type: string
          example: 0123456789
        mobile:
          type: string
          example: 0612345678
        siteWeb:
          type: string
          example: https://www.technosoft.fr
        siret:
          type: string
          example: 12345678901234
        tvaIntra:
          type: string
          example: FR12345678901
        typeClientId:
          type: string
          format: uuid
        secteurActiviteId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PROSPECT, ACTIF, INACTIF, SUSPENDU, ARCHIVE, LEAD_CHAUD, LEAD_FROID]
          default: PROSPECT
        priorite:
          type: string
          enum: [BASSE, MOYENNE, HAUTE, CRITIQUE]
          default: MOYENNE
        source:
          type: string
          enum: [PROSPECTION, RECOMMANDATION, PARTENAIRE, SALON, SITE_WEB, RESEAUX_SOCIAUX, CAMPAGNE_EMAIL, APPEL_ENTRANT, BASE_ACHETEE, AUTRE]
        chiffreAffaireAnnuel:
          type: number
          format: decimal
          example: 2500000
        effectif:
          type: integer
          example: 45
        prospectId:
          type: string
          format: uuid
        commercialId:
          type: string
          format: uuid
...
    ClientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          example: CLI-202401-0001
        nom:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
        priorite:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ClientsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ClientResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
              properties:
                total:
                  type: integer
                  example: 150
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                totalPages:
                  type: integer
                  example: 15
...
    CreateContactRequest:
      type: object
      required:
        - clientId
        - nom
        - prenom
      properties:
        clientId:
          type: string
          format: uuid
        civilite:
          type: string
          enum: [M., Mme, Mlle]
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        telephone:
          type: string
        mobile:
          type: string
        poste:
          type: string
        departement:
          type: string
        type:
          type: string
          enum: [COMMERCIAL, TECHNIQUE, COMPTABLE, DIRECTION, SUPPORT, AUTRE]
          default: COMMERCIAL
        principal:
          type: boolean
          default: false
        dateNaissance:
          type: string
          format: date
...
    ContactResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        telephone:
          type: string
        poste:
          type: string
        type:
          type: string
        statut:
          type: string
        principal:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ContactsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContactResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
              properties:
                total:
                  type: integer
                page:
                  type: integer
                limit:
                  type: integer
                totalPages:
                  type: integer

    # ==================== CONTRATS SCHEMAS ====================
    CreateContratRequest:
      type: object
      required:
        - clientId
        - titre
        - typeContrat
        - dateDebut
        - montantHT
      properties:
        clientId:
          type: string
          format: uuid
        titre:
          type: string
        description:
          type: string
        typeContrat:
          type: string
          enum: [MAINTENANCE, SERVICE, PRODUIT, PARTENARIAT, ABONNEMENT, FORFAIT, CONSULTING, FORMATION, LICENCE, SAAS]
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        montantHT:
          type: number
          format: decimal
        tauxTVA:
          type: number
          format: decimal
          default: 20.00
        devise:
          type: string
          default: EUR
        status:
          type: string
          enum: [BROUILLON, EN_ATTENTE_SIGNATURE, ACTIF, SUSPENDU, RESILIE, TERMINE, EN_RENOUVELLEMENT]
          default: BROUILLON
        estRenouvelable:
          type: boolean
          default: false
...
    ContratResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        numeroContrat:
          type: string
        titre:
          type: string
        typeContrat:
          type: string
        status:
          type: string
        montantTTC:
          type: number
          format: decimal
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    ContratsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContratResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
...
    CreateAdresseRequest:
      type: object
      required:
        - clientId
        - typeAdresse
        - ligne1
        - codePostal
        - ville
      properties:
        clientId:
          type: string
          format: uuid
        typeAdresse:
          type: string
          enum: [FACTURATION, LIVRAISON, SIEGE_SOCIAL, ETABLISSEMENT, CORRESPONDANCE]
          default: FACTURATION
        nomAdresse:
          type: string
        ligne1:
          type: string
        ligne2:
          type: string
        ligne3:
          type: string
        codePostal:
          type: string
        ville:
          type: string
        region:
          type: string
        pays:
          type: string
          default: cote d\'ivoire
        isPrincipal:
          type: boolean
          default: false
...
    AdresseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        typeAdresse:
          type: string
        ligne1:
          type: string
        codePostal:
          type: string
        ville:
          type: string
        pays:
          type: string
        isPrincipal:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdressesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdresseResponse'

    # ==================== INTERACTIONS SCHEMAS ====================
    CreateInteractionRequest:
      type: object
      required:
        - clientId
        - type
        - canal
        - sujet
      properties:
        clientId:
          type: string
          format: uuid
        contactId:
          type: string
          format: uuid
        type:
          type: string
          enum: [APPEL, EMAIL, REUNION, VISITE, SUPPORT, COMMERCIAL, TECHNIQUE, FORMATION, DEMONSTRATION, PRESENTATION, NEGOCIATION]
        canal:
          type: string
          enum: [TELEPHONE, EMAIL, EN_PERSONNE, VIDEO, CHAT, RESEAUX_SOCIAUX, PORTAL_CLIENT, MOBILE]
        sujet:
          type: string
        description:
          type: string
        dateInteraction:
          type: string
          format: date-time
        dureeMinutes:
          type: integer
        resultat:
          type: string
          enum: [POSITIF, NEUTRE, NEGATIF, A_RELANCER, A_SUIVRE, TERMINE, REPORTE, ANNULE]
        actionRequise:
          type: string
...
    InteractionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        canal:
          type: string
        sujet:
          type: string
        dateInteraction:
          type: string
          format: date-time
        resultat:
          type: string
        createdAt:
          type: string
          format: date-time
...
    InteractionsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/InteractionResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
...
    CreateDocumentRequest:
      type: object
      required:
        - clientId
        - typeDocument
        - nomFichier
        - chemin
        - taille
        - mimeType
      properties:
        clientId:
          type: string
          format: uuid
        contratId:
          type: string
          format: uuid
        typeDocument:
          type: string
          enum: [CONTRAT, FACTURE, DEVIS, AVENANT, KYC, LEGAL, TECHNIQUE, COMMERCIAL, ADMINISTRATIF, FINANCIER, RAPPORT]
        nomFichier:
          type: string
        chemin:
          type: string
        taille:
          type: integer
        mimeType:
          type: string
        description:
          type: string
        estValide:
          type: boolean
          default: true
        confidential:
          type: boolean
          default: false
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        typeDocument:
          type: string
        nomFichier:
          type: string
        taille:
          type: integer
        mimeType:
          type: string
        estValide:
          type: boolean
        dateUpload:
          type: string
          format: date-time
        confidential:
          type: boolean
    DocumentsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
...
    CreateOpportuniteRequest:
      type: object
      required:
        - clientId
        - nom
        - montantEstime
      properties:
        clientId:
          type: string
          format: uuid
        nom:
          type: string
        description:
          type: string
        montantEstime:
          type: number
          format: decimal
        probabilite:
          type: integer
          minimum: 0
          maximum: 100
          default: 50
        dateFermetureEstimee:
          type: string
          format: date-time
        etape:
          type: string
          enum: [PROSPECTION, QUALIFICATION, PROPOSITION, NEGOCIATION, FINALISATION]
          default: PROSPECTION
        statut:
          type: string
          enum: [OUVERTE, GAGNEE, PERDUE, MISE_EN_ATTENTE]
          default: OUVERTE
        source:
          type: string
...
    OpportunitesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/OpportuniteResponse'
        meta:
          type: object
          properties:
            pagination:
              type: object
            pipelineValue:
              type: number
              format: decimal
...